+++ BALSAMIC/commands/config/pon.py
@@ -8,7 +8,6 @@ import click
 
 from BALSAMIC import __version__ as balsamic_version
 from BALSAMIC.commands.options import (
-    OPTION_ADAPTER_TRIM,
     OPTION_ANALYSIS_DIR,
     OPTION_BALSAMIC_CACHE,
     OPTION_CACHE_VERSION,
@@ -16,17 +15,20 @@ from BALSAMIC.commands.options import (
     OPTION_FASTQ_PATH,
     OPTION_GENOME_INTERVAL,
     OPTION_GENOME_VERSION,
+    OPTION_SENTIEON_INSTALL_DIR,
+    OPTION_SENTIEON_LICENSE,
     OPTION_PANEL_BED,
     OPTION_PON_VERSION,
     OPTION_PON_WORKFLOW,
-    OPTION_QUALITY_TRIM,
-    OPTION_UMI,
-    OPTION_UMI_TRIM_LENGTH,
 )
 from BALSAMIC.constants.analysis import BIOINFO_TOOL_ENV, PONWorkflow
 from BALSAMIC.constants.cache import GenomeVersion
 from BALSAMIC.constants.constants import FileType
-from BALSAMIC.constants.paths import CONTAINERS_DIR
+from BALSAMIC.constants.paths import (
+    CONTAINERS_DIR,
+    SENTIEON_DNASCOPE_MODEL,
+    SENTIEON_TNSCOPE_MODEL,
+)
 from BALSAMIC.models.config import ConfigModel
 from BALSAMIC.utils.cli import (
     generate_graph,
@@ -41,7 +43,6 @@ LOG = logging.getLogger(__name__)
 
 
 @click.command("pon", short_help="Create a sample config file for PON analysis")
-@OPTION_ADAPTER_TRIM
 @OPTION_ANALYSIS_DIR
 @OPTION_BALSAMIC_CACHE
 @OPTION_CACHE_VERSION
@@ -49,16 +50,14 @@ LOG = logging.getLogger(__name__)
 @OPTION_FASTQ_PATH
 @OPTION_GENOME_VERSION
 @OPTION_GENOME_INTERVAL
+@OPTION_SENTIEON_INSTALL_DIR
+@OPTION_SENTIEON_LICENSE
 @OPTION_PANEL_BED
 @OPTION_PON_WORKFLOW
 @OPTION_PON_VERSION
-@OPTION_QUALITY_TRIM
-@OPTION_UMI
-@OPTION_UMI_TRIM_LENGTH
 @click.pass_context
 def pon_config(
     context: click.Context,
-    adapter_trim: bool,
     analysis_dir: Path,
     balsamic_cache: Path,
     cache_version: str,
@@ -66,11 +65,10 @@ def pon_config(
     fastq_path: Path,
     genome_version: GenomeVersion,
     genome_interval: Path,
+    sentieon_install_dir: Path,
+    sentieon_license: str,
     panel_bed: Path,
     pon_workflow: PONWorkflow,
-    quality_trim: bool,
-    umi: bool,
-    umi_trim_length: bool,
     version: str,
 ):
     references_path: Path = Path(balsamic_cache, cache_version, genome_version)
@@ -105,11 +103,12 @@ def pon_config(
         directory.mkdir(exist_ok=True)
 
     config_collection_dict = ConfigModel(
-        QC={
-            "adapter_trim": adapter_trim,
-            "quality_trim": quality_trim,
-            "umi_trim": umi if panel_bed else False,
-            "umi_trim_length": umi_trim_length,
+        sentieon={
+            "sentieon_install_dir": sentieon_install_dir,
+            "sentieon_license": sentieon_license,
+            "sentieon_exec": Path(sentieon_install_dir, "bin", "sentieon").as_posix(),
+            "dnascope_model": SENTIEON_DNASCOPE_MODEL.as_posix(),
+            "tnscope_model": SENTIEON_TNSCOPE_MODEL.as_posix(),
         },
         analysis={
             "case_id": case_id,