+++ BALSAMIC/constants/workflow_params.py
@@ -118,9 +118,13 @@ VCF_DICT = {
 SLEEP_BEFORE_START = 600
 
 WORKFLOW_PARAMS = {
+    "bam_post_processing": {
+        "manta_max_base_quality": 70,
+    },
     "common": {
+        "header_per_lane": "'@RG\\tID:{fastq_pattern}\\tSM:{sample_type}\\tPL:ILLUMINAi'",
+        "header_per_sample": "'@RG\\tID:{sample}\\tSM:{sample_type}\\tPL:ILLUMINAi'",
         "pcr_model": "NONE",
-        "align_header": "'@RG\\tID:{sample}\\tSM:{sample}\\tPL:ILLUMINAi'",
         "min_mapq": "20",
         "picard_fixmate": " ".join(
             [
@@ -155,9 +159,21 @@ WORKFLOW_PARAMS = {
             ]
         ),
     },
+    "insert_size_metrics": {
+        "min_read_ratio": 0.01,
+    },
     "manta": {
         "wgs_settings": "",
         "tga_settings": "--exome",
+    },
+    "mosdepth": {
+        "mapq": 20,
+        "samflag": 1796,
+        "quantize": "0:1:50:150:",
+    },
+    "sentieon_wgs_metrics": {
+        "min_base_qual": 10,
+        "cov_threshold": [50, 100, 150, 200, 250],
     },
     "vardict": {
         "allelic_frequency": "0.001",
@@ -169,7 +185,6 @@ WORKFLOW_PARAMS = {
         "vep_filters": "--compress_output bgzip --vcf --everything --hgvsg --allow_non_variant --dont_skip --buffer_size 30000 --max_sv_size 249250621 --format vcf --offline --variant_class --merged --cache --verbose --force_overwrite"
     },
     "umicommon": {
-        "align_header": "'@RG\\tID:{sample}\\tSM:{sample}\\tLB:TargetPanel\\tPL:ILLUMINA'",
         "align_intbases": 1000000,
         "filter_tumor_af": 0.0005,
     },