+++ BALSAMIC/models/config.py
@@ -178,6 +178,23 @@ class CustomFilters(BaseModel):
     umi_min_reads: str | None = None
 
 
+class Sentieon(BaseModel):
+    """
+    Class providing common functions and variables for different balsamic workflows.
+
+    Attributes:
+        sentieon_install_dir: Field(required); path to Sentieon installation directory
+        sentieon_exec:  Field(required); path to Sentieon executeable
+        sentieon_license: Field(required); Sentieon license string
+    """
+
+    sentieon_install_dir: Annotated[str, AfterValidator(is_dir)]
+    sentieon_exec: Annotated[str, AfterValidator(is_file)]
+    sentieon_license: str
+    dnascope_model: Annotated[str, AfterValidator(is_file)]
+    tnscope_model: Annotated[str, AfterValidator(is_file)]
+
+
 class ConfigModel(BaseModel):
     """
     Class providing common functions and variables for different balsamic workflows.
@@ -194,6 +211,7 @@ class ConfigModel(BaseModel):
         background_variants: Field(Path(optional)); path to BACKGROUND VARIANTS for UMI
         analysis: Field(AnalysisModel); Pydantic model containing workflow variables
         custom_filters: Field(CustomFilters); custom parameters for variant filtering
+        sentieon: Field(required); Sentieon model attributes
 
     This class also contains functions that help retrieve sample and file information,
     facilitating BALSAMIC run operations in Snakemake.
@@ -209,7 +227,7 @@ class ConfigModel(BaseModel):
         - get_final_bam_name: Return final bam name for downstream analysis.
     """
 
-    QC: QCModel
+    QC: QCModel = QCModel()
     samples: List[SampleInstanceModel]
     reference: Dict[str, Path]
     singularity: Dict[str, str]
@@ -220,6 +238,7 @@ class ConfigModel(BaseModel):
     background_variants: Optional[str] = None
     analysis: AnalysisModel
     custom_filters: CustomFilters | None = None
+    sentieon: Sentieon
 
     @field_validator("reference")
     def abspath_as_str(cls, reference: Dict[str, Path]):
@@ -380,18 +399,23 @@ class ConfigModel(BaseModel):
         bam_names = []
         for sample in self.samples:
             if sample.name == sample_name:
+                sample_type = self.get_sample_type_by_name(sample_name)
                 bam_names.extend(
                     [
-                        f"{bam_dir}{sample_name}_align_sort_{fastq_pattern}.bam"
+                        f"{bam_dir}{sample_type}.{sample_name}.{fastq_pattern}.align_sort.bam"
                         for fastq_pattern in sample.fastq_info
                     ]
                 )
         return bam_names
 
     def get_final_bam_name(
-        self, bam_dir: str, sample_name: str = None, sample_type: str = None
+        self,
+        bam_dir: str,
+        sample_name: str = None,
+        sample_type: str = None,
+        specified_suffix: str = None,
     ) -> str:
-        """Return final bam name to be used in downstream analysis."""
+        """Return bam name to be used in downstream analysis."""
 
         if not sample_name and not sample_type:
             raise ValueError(
@@ -415,10 +439,13 @@ class ConfigModel(BaseModel):
             final_bam_suffix = "dedup"
         elif self.analysis.sequencing_type == SequencingType.TARGETED:
             # Only dedup is necessary for TGA
-            final_bam_suffix = "dedup_sorted"
+            final_bam_suffix = "dedup.fixmate"
         else:
             # For WGS the bamfiles are realigned
             final_bam_suffix = "dedup.realign"
+
+        if specified_suffix:
+            final_bam_suffix = specified_suffix
 
         return f"{bam_dir}{sample_type}.{sample_name}.{final_bam_suffix}.bam"