+++ BALSAMIC/models/params.py
@@ -1,7 +1,7 @@
 """Balsamic analysis parameters models."""
 from typing import Optional
 
-from pydantic import BaseModel, ConfigDict
+from pydantic import BaseModel, ConfigDict, field_validator
 from BALSAMIC.constants.analysis import SequencingType
 
 
@@ -10,19 +10,27 @@ class ParamsCommon(BaseModel):
 
     Attributes:
         pcr_model: str (required). PCR indel model used to weed out false positive indels. Eg: none- PCR free samples.
-        align_header: str (required); header line appended to the aligned BAM output
         min_mapq: int (required); minimum mapping quality score. Eg: 20- probability of mapping random read at 99% accuracy
         picard_fixmate: str (required), fix read mate information in bam file
         picard_RG_normal: str (required); replace readgroups in normal bam file
         picard_RG_tumor: str (required); replace readgroups in tumor bam file
     """
 
-    align_header: str
     pcr_model: str
     min_mapq: int
     picard_fixmate: str
     picard_RG_normal: str
     picard_RG_tumor: str
+
+
+class ParamsInsertSizeMetrics(BaseModel):
+    """This class defines the common params settings used for the InsertSizeMetricsAlgo
+
+    Attributes:
+        min_read_ratio: float(required). Minimum ratio of reads for a read category to be included in the output histogram
+    """
+
+    min_read_ratio: float
 
 
 class ParamsManta(BaseModel):
@@ -35,6 +43,39 @@ class ParamsManta(BaseModel):
 
     wgs_settings: str
     tga_settings: str
+
+
+class ParamsMosdepth(BaseModel):
+    """This class defines the params settings used as constants in Mosdepth rule.
+
+    Attributes:
+        mapq: str(required); mapping quality threshold, reads with a quality less than this value are ignored
+        samflag: str(required); exclude reads with any of the bits in FLAG set
+        quantize: str(required); merges adjacent bases as long as they fall in the same coverage bins e.g. (10-20)
+    """
+
+    mapq: int
+    samflag: int
+    quantize: str
+
+
+class ParamsSentieonWGSMetrics(BaseModel):
+    """This class defines the params settings used as constants in Sentieon WGS Metrics rule.
+
+    Attributes:
+        min_base_qual: int(required); base quality threshold, bases with a quality less than this value are ignored
+        cov_threshold: list(required); coverage threshold list
+    """
+
+    min_base_qual: int
+    cov_threshold: str
+
+    @field_validator("cov_threshold", mode="before")
+    def parse_into_arguments(cls, cov_threshold):
+        param_values = []
+        for value in cov_threshold:
+            param_values.append(f"--cov_thresh {value}")
+        return " ".join(param_values)
 
 
 class ParamsVardict(BaseModel):
@@ -66,13 +107,8 @@ class ParamsVEP(BaseModel):
 class QCModel(BaseModel):
     """Contains settings for quality control and pre-processing
     Attributes:
-        picard_rmdup : Field(bool); whether duplicate removal is to be applied in the workflow
         adapter : Field(str(AATGATACGGCGACCACCGAGATCTACACTCTTTCCCTACACGACGCTCTTCCGATCT)); adapter sequence to trim
-        quality_trim : Field(bool); whether quality trimming it to be performed in the workflow
-        adapter_trim : Field(bool); whether adapter trimming is to be performed in the workflow
-        umi_trim : Field(bool); whether UMI trimming is to be performed in the workflow
         min_seq_length : Field(str(int)); minimum sequence length cutoff for reads
-        umi_trim_length : Field(str(int)); length of UMI to be trimmed from reads
         n_base_limit : Field(str(int)); supports filtering by limiting the N base number
 
     Raises:
@@ -83,13 +119,8 @@ class QCModel(BaseModel):
     """
 
     model_config = ConfigDict(coerce_numbers_to_str=True)
-    picard_rmdup: bool = False
     adapter: str = "AATGATACGGCGACCACCGAGATCTACACTCTTTCCCTACACGACGCTCTTCCGATCT"
-    quality_trim: bool = True
-    adapter_trim: bool = False
-    umi_trim: bool = False
     min_seq_length: str = "25"
-    umi_trim_length: str = "5"
     n_base_limit: str = "50"
 
 
@@ -98,12 +129,10 @@ class UMIParamsCommon(BaseModel):
 
     Attributes:
         align_format: str (required); output alignment format. eg. 'BAM'
-        align_header: str (required); header line appended to the aligned BAM output
         align_intbases: int; input bases in each batch regardless of threads, for reproducibility
         filter_tumor_af: float (required); settings to filter minimum allelic frequency
     """
 
-    align_header: str
     align_intbases: int
     filter_tumor_af: float
 
@@ -123,7 +152,7 @@ class UMIParamsConsensuscall(BaseModel):
 
     Attributes:
         align_format: str (required); output alignment format. eg. 'BAM'
-            filter_minreads: str (required); settings to filter consensus tags based on group size
+        filter_minreads: str (required); settings to filter consensus tags based on group size
         tag: str; Logic UMI tag
     """
 
@@ -155,6 +184,16 @@ class UMIParamsTNscope(BaseModel):
     disable_detect: str
 
 
+class BAMPostProcessingParams(BaseModel):
+    """This class defines the params settings used as constants bam post processing rules
+
+    Attributes:
+       manta_max_base_quality: int (required); the maximum base quality in bamfile used downstream in Manta rules
+    """
+
+    manta_max_base_quality: int
+
+
 class BalsamicWorkflowConfig(BaseModel):
     """Defines set of rules in balsamic workflow
 
@@ -162,7 +201,9 @@ class BalsamicWorkflowConfig(BaseModel):
 
     Attributes:
         common: global params defined across all rules in balsamic workflow
+        bam_post_processing: params used in bam post-processing rules
         manta: params used in the manta rules
+        mosdepth: params used in mosdepth rule
         umicommon: global params defined across specific rules in UMI workflow
         vep: global params defined in the rule vep
         vardict: params defined in the rule vardict
@@ -174,8 +215,12 @@ class BalsamicWorkflowConfig(BaseModel):
         - get_manta_settings: Return setting for manta rule
     """
 
+    bam_post_processing: BAMPostProcessingParams
     common: ParamsCommon
+    insert_size_metrics: ParamsInsertSizeMetrics
     manta: ParamsManta
+    mosdepth: ParamsMosdepth
+    sentieon_wgs_metrics: ParamsSentieonWGSMetrics
     vardict: ParamsVardict
     vep: ParamsVEP
     umicommon: UMIParamsCommon